requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.6.5
django-bootstrap-pagination==1.5.1
django-bootstrap3==5.2.0
django-classy-tags==0.5.1
django-filebrowser==3.5.6
django-grappelli==2.5.3
-e git://github.com/django-grappelli-extensions/django-grappelli-extensions.git@bca8e7542c78d6d8a63b8b3906c40a1160db55fc#egg=django_grappelli_extensions-dev
django-taggit==0.12.1
django-taggit-autosuggest==0.2.6
django-tinymce==1.5.2
gunicorn==18.0
Markdown==2.4.1
Pillow==2.4.0
psycopg2==2.5.3
pystache==0.5.4
python-decouple==2.2
pytz==2014.2
sorl-thumbnail==12.2
South==0.8.4
static==1.0.2
static3==0.5.1
Unipath==1.0
